// src/app/shared/helpers/enums/catalog/cat.estatus-contacto-telefonico.enum.ts
//FYI: Cada valor representa la clave del catálogo
export enum ECatEstatusContactoTelefonico {
  DATOS_DE_CONTACTO_INCORRECTOS = 1,
  NO_CONTESTA = 2,
  EN_ESPERA_DE_DOCUMENTOS_DIGITALES = 3,
  NO_DESEA_CONTINUAR_CON_EL_TRAMITE = 4,
  EN_ESPERA_DE_FIRMA_CLIENTE = 5,
  EN_ESPERA_DE_FIRMA_EJECUTIVO = 6,
  VALIDANDO_FIRMAS = 7,
  FORMATOS_ENVIADOS_A_ASEGURADORA = 8,
  EN_ESPERA_DE_GENERACION_DE_FORMATOS = 9,
  EN_ESPERA_DE_VALIDACION_AFIANZADORA = 10,
  DOCUMENTACION_FISICA_SOLICITADA = 11,
  EN_ESPERA_DE_VALIDACION_DE_ORIGINALES = 12,
  TELEFONO_INCORRECTO = 13,
  TELEFONO_SUSPENDIDO = 14,
  SE_CONTACTO_CLIENTE = 15,
}